pyproject.toml
[tool.poetry]
name = "Contacts_web"
version = "0.1.0"
description = ""
authors = ["shuguruiua <popov.yk@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "0.110.0"
sqlalchemy = "2.0.28"
alembic = "1.13.1"
asyncpg = "0.29.0"
uvicorn = "0.29.0"
pydantic = {extras = ["email"], version = "2.6.4"}
python-multipart = "0.0.9"
passlib = {extras = ["bcrypt"], version = "1.7.4"}
libgravatar = "1.0.4"
python-jose = {extras = ["cryptography"], version = "3.3.0"}
bcrypt = "4.0.01"
python-dotenv = "^1.0.1"
fastapi-mail = "^1.4.1"
fastapi-limiter = "^0.1.6"
cloudinary = "^1.39.1"
pytest = "^8.1.1"

[tool.poetry.group.dev.dependencies]
sphinx = "7.2.6"

[tool.poetry.group.test.dependencies]
aiosqlite = "0.18.0"
pytest_asyncio = "0.21.0"
httpx = "0.23.3"
pytest-cov = "4.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "--doctest-modules"
testpaths = [
    "tests", ]
pythonpath = "."
filterwarnings = "ignore::DeprecationWarning"